Add Monster Book UI, idle HP/MP regen, and heal floating numbers
Monster Book: card grid with mob sprites, detail overlay with animated
sprite (stand/move/die), HP/MP gauge bars from status bar, drop items,
and stat display. Category tabs, search, and page navigation.
Note: Monster Book UI is broken but functional — positioning needs work.

Idle regen: HEAL_OVER_TIME packet (opcode 0x59) sent every ~10s while
idle. HP = level/5+2, MP = level/5+INT/20+3, mage MP Recovery skill
bonus (Level*SkillLevel/10), 3x chair bonus. Pauses during attack,
hit stun, invincibility, and death.

Heal numbers: blue floating numbers via new DamageNumber::HEAL type
shown when HP increases from regen.

+++ src/Character/MonsterBook.h
@@ -29,11 +29,11 @@ namespace ms
 		MonsterBook();
 
 		void set_cover(int32_t cov);
-		void add_card(int16_t cardid, int8_t level);
+		void add_card(int32_t cardid, int8_t level);
 
 		int32_t get_cover() const;
-		int8_t get_card_level(int16_t cardid) const;
-		const std::map<int16_t, int8_t>& get_cards() const;
+		int8_t get_card_level(int32_t cardid) const;
+		const std::map<int32_t, int8_t>& get_cards() const;
 		int16_t get_total_cards() const;
 		int32_t get_book_level() const;
 		int32_t get_normal_cards() const;
@@ -41,6 +41,6 @@ namespace ms
 
 	private:
 		int32_t cover;
-		std::map<int16_t, int8_t> cards;
+		std::map<int32_t, int8_t> cards;
 	};
 }
